+++ apps/core-rest-api/src/app/adapters/controllers/api/api.module.ts
@@ -25,6 +25,8 @@ import { NestjsCreateAppointmentService } from './use-cases/appointment/create-a
 import { NestjsCreateClinicService } from './use-cases/clinic/create-clinic/nestjs-create-clinic.service';
 import { NestjsDeleteClinicService } from './use-cases/clinic/delete-clinic/nestjs-delete-clinic.service';
 import { NestjsUpdateClinicService } from './use-cases/clinic/update-clinic/nestjs-update-clinic.service';
+import { CreatePatientAppointmentRegistryController } from './use-cases/patient-appointment-registry/create-patient-appointment-registry/create-patient-appointment-registry.controller';
+import { NestjsCreatePatientAppointmentRegistryService } from './use-cases/patient-appointment-registry/create-patient-appointment-registry/nestjs-create-patient-appointment-registry.service';
 import { NestjsCreatePatientService } from './use-cases/patient/create-patient/nestjs-create-patient.service';
 import { NestjsDeletePatientService } from './use-cases/patient/delete-patient/nestjs-delete-patient.service';
 import { NestjsAuthenticatePsychologistService } from './use-cases/psychologist/authenticate-psychologist/nestjs-authenticate-psychologist.service';
@@ -53,6 +55,7 @@ import { NestjsUpdatePsychologistService } from './use-cases/psychologist/update
     DeleteClinicController,
     CreatePatientController,
     DeletePatientController,
+    CreatePatientAppointmentRegistryController,
     CreateAppointmentController
   ],
   providers: [
@@ -67,6 +70,7 @@ import { NestjsUpdatePsychologistService } from './use-cases/psychologist/update
     NestjsDeleteClinicService,
     NestjsCreatePatientService,
     NestjsDeletePatientService,
+    NestjsCreatePatientAppointmentRegistryService,
     NestjsCreateAppointmentService
   ],
 })