+++ apps/core-rest-api/src/app/adapters/database/repositories/repositories.module.ts
@@ -5,9 +5,11 @@ import { PatientDatabaseRepository } from '../../../core/domains/patient/reposit
 import { PsychologistDatabaseRepository } from '../../../core/domains/psychologist/repositories/database-repository';
 
 import { AppointmentDatabaseRepository } from '../../../core/domains/appointment/repositories/database-repository';
+import { PatientAppointmentRegistryDatabaseRepository } from '../../../core/domains/patient-appointment-registry/repositories/database-repository';
 import { PostgreSqlPrismaOrmService } from '../infra/prisma/prisma.service';
 import { PostgresqlPrismaOrmAppointmentRepository } from './appointment/postgres-prisma-orm-appointment-repository';
 import { PostgresqlPrismaOrmClinicRepository } from './clinic/postgres-prisma-orm-clinic-repository';
+import { PostgresqlPrismaOrmPatientAppointmentRegistryRepository } from './patient-appointment-registry/postgresql-prisma-orm-registry-repository';
 import { PostgresqlPrismaOrmPatientRepository } from './patient/postgres-prisma-orm-patient-repository';
 import { PostgresqlPrismaOrmPsychologistRepository } from './psychologist/postgresql-prisma-orm-psychologist-repository';
 
@@ -29,6 +31,10 @@ import { PostgresqlPrismaOrmPsychologistRepository } from './psychologist/postgr
       useClass: PostgresqlPrismaOrmPatientRepository,
     },
     {
+      provide: PatientAppointmentRegistryDatabaseRepository,
+      useClass: PostgresqlPrismaOrmPatientAppointmentRegistryRepository,
+    },
+    {
       provide: AppointmentDatabaseRepository,
       useClass: PostgresqlPrismaOrmAppointmentRepository,
     },
@@ -46,6 +52,10 @@ import { PostgresqlPrismaOrmPsychologistRepository } from './psychologist/postgr
     {
       provide: PatientDatabaseRepository,
       useClass: PostgresqlPrismaOrmPatientRepository,
+    },
+    {
+      provide: PatientAppointmentRegistryDatabaseRepository,
+      useClass: PostgresqlPrismaOrmPatientAppointmentRegistryRepository,
     },
     {
       provide: AppointmentDatabaseRepository,