+++ apps/core-rest-api/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.0-beta.1](https://github.com/ItaloRAmaral/cliniccontrol/compare/@cliniccontrol/core-rest-api@1.4.0-beta.0...@cliniccontrol/core-rest-api@1.4.0-beta.1) (2024-02-04)
+
+### Features
+
+- add mapper for registry ([e0d1247](https://github.com/ItaloRAmaral/cliniccontrol/commit/e0d1247dae8996baefaeb5fc0cbb6146b3916885))
+- **CC-118:** add controllers and service in api module ([452c503](https://github.com/ItaloRAmaral/cliniccontrol/commit/452c503e0f063855dc707c7d29d866eb2ecd438f))
+- **CC-118:** add dto to appointment registry ([44dcdee](https://github.com/ItaloRAmaral/cliniccontrol/commit/44dcdeee5ff32e50e2eddec02b134f6cc4d4b000))
+- **CC-118:** add methods in prisma registry repository ([b944e02](https://github.com/ItaloRAmaral/cliniccontrol/commit/b944e0229ac910ea7be8e4f26628c18640eb66fd))
+- **CC-118:** add registry fake client ([1615baa](https://github.com/ItaloRAmaral/cliniccontrol/commit/1615baa02af1250c893fdc76a6d2f53c55db75d7))
+- **CC-118:** add registry mapper ([3f69213](https://github.com/ItaloRAmaral/cliniccontrol/commit/3f69213347e77136a95477a2e80ea300ba266ff3))
+- **CC-118:** add repository in repositories module ([8142178](https://github.com/ItaloRAmaral/cliniccontrol/commit/8142178c2331615f0caf54677431ba32bc5c34e1))
+- **CC-118:** add test e2e to appointment registry ([f4a4d78](https://github.com/ItaloRAmaral/cliniccontrol/commit/f4a4d78886a762bbd6f2033d0b495a4c8fdbec05))
+- **CC-118:** create controller to appointment registry ([f11b910](https://github.com/ItaloRAmaral/cliniccontrol/commit/f11b9105394f6e202ac5c7ad631e882b9e311f8a))
+- **CC-118:** create fake client to registry ([a547515](https://github.com/ItaloRAmaral/cliniccontrol/commit/a547515207a3dfad0968eefb21b581f56b9c3d14))
+- **CC-118:** create service to appointment registry ([0b39908](https://github.com/ItaloRAmaral/cliniccontrol/commit/0b399083a9bf7478fa8f8f64961749d098092720))
+- create PostgresqlPrismaOrmRegistryRepository ([cdbc7de](https://github.com/ItaloRAmaral/cliniccontrol/commit/cdbc7de00a77efe081c2d3b9082ad8a3fce481cf))
+
 ## [1.4.0-beta.0](https://github.com/ItaloRAmaral/cliniccontrol/compare/@cliniccontrol/core-rest-api@1.3.0...@cliniccontrol/core-rest-api@1.4.0-beta.0) (2024-01-12)
 
 ### Features